requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
coverage==5.1
distlib==0.3.0
Django==3.0.5
django-factory-boy==1.0.0
django-filter==2.2.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.3
filelock==3.0.12
importlib-metadata==1.6.0
importlib-resources==1.5.0
inflection==0.4.0
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
virtualenv==20.0.18
wcwidth==0.1.9
zipp==3.1.0
